Pull up similar functionality from resource bundle & preload processors
